Add theme customization, region filter, and proxy updates (#74)
* feat(proxy): add Masque option to proxy settings

* feat(access-control): add region quick select feature for app filtering (#38)

Introduce a new dropdown for selecting apps based on region (China or Overseas).
This feature enhances user experience by allowing quick selection of apps
from specific regions, improving the accessibility of app management.

Also, refactor related ViewModel logic to support the new functionality.

* feat(theme): add theme seed color customization feature

Introduce the ability to customize the theme seed color in the app settings.
This allows users to select a custom color for the theme, enhancing personalization
and user experience. The implementation includes a color picker and options to
reset to default values.

* bump(version): update project version to 0.3.8